Introduces `search` feature enabled by default
This change moves the filesystem search functionality behind a feature
flag so that consuming libraries can opt-out of glob if they do not need
it.

For example, users using `include_dir` to serve embedded assets for a game
or a webserver likely know the exact paths of the files they want to
serve.

1. The `search` feature is enabled by default for backwards compatibility.
2. `Dir.find`, the `globs` module, and the `glob` dependency are now
   gated behind the `search` feature.
3. `glob` is now an optional dependency
4. This change also updates the CI configuration to test all features in
   case new ones are introduced in the future
