+++ appveyor.yml
@@ -28,10 +28,10 @@ install:
 test_script:
   # we don't run the "test phase" when doing deploys
   - if [%APPVEYOR_REPO_TAG%]==[false] (
-      cargo build --target %TARGET% &&
-      cargo build --target %TARGET% --release &&
-      cargo test --target %TARGET% &&
-      cargo test --target %TARGET% --release
+      cargo build --workspace --all-features --target %TARGET% &&
+      cargo build --workspace --all-features --target %TARGET% --release &&
+      cargo test --workspace --all-features --target %TARGET% &&
+      cargo test --workspace --all-features --target %TARGET% --release
     )
 
 cache:
