+++ include_dir/Cargo.toml
@@ -23,10 +23,11 @@ branch = "master"
 repository = "Michael-F-Bryan/include_dir"
 
 [dependencies]
-glob = "0.3"
+glob = { version = "0.3", optional = true }
 proc-macro-hack = "0.5"
 include_dir_impl = { path = "../include_dir_impl", version = "0.5.1-alpha.0" }
 
 [features]
-default = []
+default = [ "search" ]
 example-output = []
+search = [ "glob" ]
