+++ include_dir/tests/integration_test.rs
@@ -1,7 +1,4 @@
-#[macro_use]
-extern crate include_dir;
-
-use include_dir::Dir;
+use include_dir::{include_dir, Dir};
 use std::path::Path;
 
 const PARENT_DIR: Dir<'_> = include_dir!(".");
